% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{banquereau_2010}
\alias{banquereau_2010}
\title{Arctic Surfclam Morphometrics for Banquereau 2010 Survey}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 7074 rows and 12 columns.
}
\usage{
banquereau_2010
}
\description{
Arctic Surfclam Morphometrics for Banquereau 2010 Survey
}
\keyword{datasets}
