% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_stvonb_distributions}
\alias{get_stvonb_distributions}
\title{List of stvonb covariance functions}
\usage{
get_stvonb_distributions(which = "covariance")
}
\arguments{
\item{which}{Unused}
}
\value{
A named character vector giving the implemented covariance functions
}
\description{
Print a list of implemented covariance functions.
}
\section{Covariance Functions}{

\describe{
\item{exponential - }{Matern covariance with smoothness nu = 0.5. Results
in a non-differentiable Gaussian random field.}
\item{gaussian - }{Limiting function of the Matern covariance as smoothness
nu approaches infinity. Results in smooth (infinitely differentiable)
Gaussian random field. The variance parameter in this covariance function
gives the marginal variance of the spatial field.}
\item{matern - }{Matern covariance with arbitrary smoothness nu. Results in
Gaussian random fields that are floor(nu) times differentiable. The
smoothness parameter is hard to estimate, so fixing it to a set value is
recommended. This option will run much slower than other specific values
of the Matern covariance function (e.g. exponential).}
\item{matern32 - }{Matern covariance with smoothness nu = 1.5. Results in
a smooth (once differentiable) Gaussian random field.}
}
}

