% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grand_banks_2006_9}
\alias{grand_banks_2006_9}
\title{Arctic Surfclam Morphometrics for Grand Banks 2006-9 Survey}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 9160 rows and 12 columns.
}
\usage{
grand_banks_2006_9
}
\description{
Arctic Surfclam Morphometrics for Grand Banks 2006-9 Survey
}
\keyword{datasets}
