% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{parameters-class}
\alias{parameters-class}
\title{An S4 class to collect process and observations parameters.}
\description{
An S4 class to collect process and observations parameters.
}
\seealso{
process_parameters_class, observation_parameters_class, stvonb_class

Other stvonb_classes: 
\code{\link{TMB_out-class}},
\code{\link{dag-class}},
\code{\link{long_stars-class}},
\code{\link{observation_parameters-class}},
\code{\link{observations-class}},
\code{\link{process-class}},
\code{\link{process_parameters-class}},
\code{\link{settings-class}},
\code{\link{stvonb-class}},
\code{\link{tracing-class}}
}
\concept{stvonb_classes}
\keyword{internal}
