% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stvonb.R
\name{predict_stvonb}
\alias{predict_stvonb}
\title{Predict spatial growth curves}
\usage{
predict_stvonb(fit, raster, age, quantiles = c(0.05, 0.5, 0.95))
}
\arguments{
\item{fit}{A fitted model from fit_stvonb}

\item{raster}{A stars raster}

\item{age}{Ages to use for length-at-age predictions, if any.}

\item{quantiles}{The quantiles to predict. Defaults to a 95\% confidence interval around the mean}
}
\value{
A list with the spatial von Bertalanffy parameter predictions and spatial length-at-age predictions
}
\description{
Predict spatial growth curves
}
