% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stvonb-package.R
\docType{package}
\name{stvonb-package}
\alias{stvonb}
\alias{stvonb-package}
\title{stvonb: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
