% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/stvonb.R, R/show.R
\docType{class}
\name{stvonb-class}
\alias{stvonb-class}
\alias{settings,stvonb-method}
\alias{tracing,stvonb-method}
\alias{time_effects,stvonb-method}
\alias{time_effects<-,stvonb-method}
\alias{pg_re,stvonb-method}
\alias{pg_re<-,stvonb-method}
\alias{tg_re,stvonb-method}
\alias{tg_re<-,stvonb-method}
\alias{random_effects,stvonb-method}
\alias{persistent_graph,stvonb-method}
\alias{transient_graph,stvonb-method}
\alias{graph,stvonb-method}
\alias{covariance_function,stvonb-method}
\alias{covariance_function<-,stvonb-method}
\alias{space_parameters,stvonb-method}
\alias{space_parameters<-,stvonb-method}
\alias{time_parameters,stvonb-method}
\alias{time_parameters<-,stvonb-method}
\alias{dat,stvonb-method}
\alias{dat<-,stvonb-method}
\alias{data_predictions,stvonb-method}
\alias{data_predictions<-,stvonb-method}
\alias{response_parameters,stvonb-method}
\alias{response_parameters<-,stvonb-method}
\alias{fixed_effects,stvonb-method}
\alias{fixed_effects<-,stvonb-method}
\alias{distance_units,stvonb-method}
\alias{distance_units<-,stvonb-method}
\alias{formula,stvonb-method}
\alias{formula<-,stvonb-method}
\alias{parameters,stvonb-method}
\alias{parameters<-,stvonb,parameters-method}
\alias{convergence,stvonb-method}
\alias{timing,stvonb-method}
\alias{show,stvonb-method}
\title{An S4 class describing a stvonb model.}
\usage{
\S4method{settings}{stvonb}(x)

\S4method{tracing}{stvonb}(x)

\S4method{time_effects}{stvonb}(x)

\S4method{time_effects}{stvonb}(x) <- value

\S4method{pg_re}{stvonb}(x)

\S4method{pg_re}{stvonb}(x) <- value

\S4method{tg_re}{stvonb}(x)

\S4method{tg_re}{stvonb}(x) <- value

\S4method{random_effects}{stvonb}(x)

\S4method{persistent_graph}{stvonb}(x)

\S4method{transient_graph}{stvonb}(x)

\S4method{graph}{stvonb}(x)

\S4method{covariance_function}{stvonb}(x)

\S4method{covariance_function}{stvonb}(x) <- value

\S4method{space_parameters}{stvonb}(x)

\S4method{space_parameters}{stvonb}(x) <- value

\S4method{time_parameters}{stvonb}(x)

\S4method{time_parameters}{stvonb}(x) <- value

\S4method{dat}{stvonb}(x)

\S4method{dat}{stvonb}(x) <- value

\S4method{data_predictions}{stvonb}(x)

\S4method{data_predictions}{stvonb}(x) <- value

\S4method{response_parameters}{stvonb}(x)

\S4method{response_parameters}{stvonb}(x) <- value

\S4method{fixed_effects}{stvonb}(x)

\S4method{fixed_effects}{stvonb}(x) <- value

\S4method{distance_units}{stvonb}(x)

\S4method{distance_units}{stvonb}(x) <- value

\S4method{formula}{stvonb}(x)

\S4method{formula}{stvonb}(x) <- value

\S4method{parameters}{stvonb}(x)

\S4method{parameters}{stvonb,parameters}(x) <- value

\S4method{convergence}{stvonb}(x)

\S4method{timing}{stvonb}(x)

\S4method{show}{stvonb}(object)
}
\arguments{
\item{x}{An object}

\item{value}{A replacement value}

\item{object}{An object}
}
\description{
An S4 class describing a stvonb model.
}
\section{Functions}{
\itemize{
\item \code{settings(stvonb)}: Get model settings

\item \code{tracing(stvonb)}: Get tracing information, see \link{tracing}.

\item \code{time_effects(stvonb)}: Get temporal random effects

\item \code{time_effects(stvonb) <- value}: Set temporal random effects

\item \code{pg_re(stvonb)}: Get persistent graph random effects

\item \code{pg_re(stvonb) <- value}: Set spatio-temporal random effects

\item \code{tg_re(stvonb)}: Get transient graph random effects

\item \code{tg_re(stvonb) <- value}: Set transient graph random effects

\item \code{random_effects(stvonb)}: Get list containing the persistent graph random
effects and the transient graph random effects.

\item \code{persistent_graph(stvonb)}: Get persistent graph

\item \code{transient_graph(stvonb)}: Get transient graph

\item \code{graph(stvonb)}: Get a list containing the persistent and transient graphs

\item \code{covariance_function(stvonb)}: Get covariance function

\item \code{covariance_function(stvonb) <- value}: Set covariance function(s). Run
get_stvonb_distributions("covariance") for valid covariance functions.
Setting the covariance function also overwrites the spatial parameters.

\item \code{space_parameters(stvonb)}: Get spatial parameters

\item \code{space_parameters(stvonb) <- value}: Set spatial parameters

\item \code{time_parameters(stvonb)}: Get time parameters

\item \code{time_parameters(stvonb) <- value}: Set time parameters

\item \code{dat(stvonb)}: Get data including response variables, time indices,
locations, covariates, etc.

\item \code{dat(stvonb) <- value}: Set data. Warning: if you want to add new
rows you must give each additional row a value in the graph_idx column,
or create a new object from scratch.

\item \code{data_predictions(stvonb)}: Get data predictions, a long_stars object
with the data (see \code{dat}) and associated random effect
predictions.

\item \code{data_predictions(stvonb) <- value}: Set data predictions

\item \code{response_parameters(stvonb)}: Get response distribution parameters

\item \code{response_parameters(stvonb) <- value}: Set response distribution parameters

\item \code{fixed_effects(stvonb)}: Get fixed effect parameters

\item \code{fixed_effects(stvonb) <- value}: Set fixed effect parameters

\item \code{distance_units(stvonb)}: Get distance units used for the model

\item \code{distance_units(stvonb) <- value}: Set distance units used for the model

\item \code{formula(stvonb)}: Get formula used for model

\item \code{formula(stvonb) <- value}: Set formula used for the model

\item \code{parameters(stvonb)}: Get all model parameters as a parameters object

\item \code{parameters(x = stvonb) <- value}: Set all model parameters with a new
parameters object

\item \code{convergence(stvonb)}: Get convergence message from optimizer

\item \code{timing(stvonb)}: Get all timing information, see \link{tracing}

\item \code{show(stvonb)}: Print method

}}
\section{Slots}{

\describe{
\item{\code{process}}{A process object.}

\item{\code{observations}}{An observations object.}

\item{\code{settings}}{A settings object.}
}}

\seealso{
stvonb_prepare

Other stvonb_classes: 
\code{\link{TMB_out-class}},
\code{\link{dag-class}},
\code{\link{long_stars-class}},
\code{\link{observation_parameters-class}},
\code{\link{observations-class}},
\code{\link{parameters-class}},
\code{\link{process-class}},
\code{\link{process_parameters-class}},
\code{\link{settings-class}},
\code{\link{tracing-class}}
}
\concept{stvonb_classes}
