% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/fit.R
\name{stvonb_fit}
\alias{stvonb_fit}
\alias{stvonb_fit,stvonb-method}
\title{Fit a model object.}
\usage{
stvonb_fit(object, ...)

\S4method{stvonb_fit}{stvonb}(object, silent = FALSE, ...)
}
\arguments{
\item{object}{A model object to be fitted.}

\item{...}{Options to be passed}

\item{silent}{Should intermediate calculations be printed?}
}
\value{
A fitted model object
}
\description{
Takes an unfitted model object and performs inference.
}
\section{Methods (by class)}{
\itemize{
\item \code{stvonb_fit(stvonb)}: Takes an unfitted stvonb object and performs
maximum likelihood inference via the \code{TMB} package to obtain parameter
and random effect estimates with their associated standard errors. Options
in the \dots argument are passed to TMB::MakeADFun and TMB::sdreport.

}}
