% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{stvonb_prepare}
\alias{stvonb_prepare}
\alias{stvonb_prepare_process}
\alias{stvonb_prepare_observations}
\title{Create an object of class \code{stvonb}.}
\usage{
stvonb_prepare(
  formula,
  data,
  mesh,
  silent = TRUE,
  distance_units = "km",
  fit = FALSE,
  ...
)

stvonb_prepare_process(data, mesh, settings)

stvonb_prepare_observations(data, process, settings)
}
\arguments{
\item{formula}{A formula object. See the 'Details' section below.}

\item{data}{An \code{sf} object containing point geometries, and any other
variables needed to fit the model.}

\item{mesh}{An INLA mesh object. See INLA::inla.mesh.2d.}

\item{silent}{Logical. Should intermediate calculations be printed?}

\item{distance_units}{Any value that can be used as a \code{units} object
from the \code{units} package. Which distance units should the model use?
Defaults to "km".}

\item{fit}{Logical (Default=FALSE). Should parameter estimates be found?
If so, the starting values for the optimizer will use the default values.}

\item{...}{Extra options to pass to \link{stvonb_fit} if fit=TRUE}

\item{settings}{A settings object}

\item{process}{A process object.}
}
\value{
An stvonb object. If fit=TRUE, the returned model parameters will be
estimated using the \link{stvonb_fit} function using the default starting
values.
}
\description{
'stvonb_prepare' is used to take an existing \verb{simple features} data.frame
with point geometries, time information, covariates, and age and length
variables and perform all of the pre-processing steps necessary to fit a
model with the \code{stvonb_fit} function. See the description for
\link{stvonb_prepare_process} and \link{stvonb_prepare_observations} for more
details on how each part is prepared.
}
\details{
The formula object should always be of the form \code{l ~ age(a) +
  x + z + time(t, type = "ar1") + space("matern", nu = 1.5)},
though possibly with some terms missing.

The \code{time(...)} term indicates which column, if any, holds the time
index. The variable t should be replaced with the desired time index. There
are currently three valid options for the `type' argument in
\code{time(t,type="ar1")} -- "ar1" for an AR(1) structure, "rw" for a
random walk, and "independent" for independent spatial fields each year. If
the \code{time(...)} term is missing, all observations are assumed to be
at the same time and a purely spatial model is used.

The \code{space(...)} term specifies the spatial covariance function. See
\code{get_starve_distributions("covariance")} for valid names to supply.
If using the "matern" option you can supply a value for the smoothness
parameter nu, which will be held constant in model fitting. If nu is not
given, then it will be freely estimated in the model. If the
\code{space(...)} term as a whole is missing, an exponential covariance
function is assumed.
}
\section{Functions}{
\itemize{
\item \code{stvonb_prepare_process()}: Creates a new process object with the correct
dimensions for the temporal random effects, persistent graph random
effects, and transient graph random effects. Initializes the temporal and
spatial parameters for the model according to the options specified in the
formula element of the settings argument. Constructs the persistent and
transient graph, see \link{construct_graph}.

Before creating the transient graph any location in data that
is present in the mesh is removed.

\item \code{stvonb_prepare_observations()}: Creates a new observation object with the correct dimensions
for the random effect predictions. Initializes the response distribution
and fixed effect parameters for the model according to the options specified
in the formula element of the settings argument. Also adds a column "graph_idx"
to the supplied data.

}}
\seealso{
stvonb_class
}
