% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survey_morphometrics}
\alias{survey_morphometrics}
\title{Arctic Surfclam Morphometrics for DFO Surveys on the Scotian Shelf}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 26478 rows and 12 columns.
}
\usage{
survey_morphometrics
}
\description{
Arctic Surfclam Morphometrics for DFO Surveys on the Scotian Shelf
}
\keyword{datasets}
