% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stvonb.R
\name{fit_stvonb}
\alias{fit_stvonb}
\title{Fit a von Bertalanffy growth curve with spatially  varying parameters.}
\usage{
fit_stvonb(data, nodes = data, ...)
}
\arguments{
\item{data}{An sf data.frame with point geometries, an age column, and a length column.}

\item{nodes}{Either an sf data.frame giving the point location of nodes or an integer giving the number of nodes which will be smartly chosen.}

\item{...}{Additional arguments, particularly those to pass to nnspline::create_nnspline, robustifyRTMB::robustly_optimize, and mcreportRTMB::mcreport.}
}
\value{
A list containing the model fit, a template spline, and mcreported spline realizations
}
\description{
Fit a von Bertalanffy growth curve with spatially  varying parameters.
}
